requirements.txt
fastapi
uvicorn
streamlit
requests
langgraph
openai
groqp
ipython
langchain